Refactor preprocesso
