+++ framework/wazuh/rbac/preprocessor.py
