+++ framework/wazuh/rbac/tests/test_preprocessor.py
@@ -7,7 +7,7 @@ from unittest.mock import patch
 import json
 import os
 import pytest
-from wazuh.rbac.pre_processor import optimize_resources
+from wazuh.rbac.preprocessor import optimize_resources
 
 
 test_path = os.path.dirname(os.path.realpath(__file__))
@@ -25,6 +25,6 @@ outputs = [test_case['processed_policies'] for test_case in file]
 @pytest.mark.parametrize('input, output', zip(inputs, outputs))
 def test_expose_resources(input, output):
 
-    with patch('wazuh.rbac.pre_processor.RBAChecker.run', return_value=input):
+    with patch('wazuh.rbac.preprocessor.RBAChecker.run', return_value=input):
         preprocessed_policies = optimize_resources()
         assert (preprocessed_policies == output)
